--- a/filter/expand_json.go
+++ b/filter/expand_json.go
@@ -7,6 +7,7 @@ import (
 	"sync/atomic"
 
 	"github.com/AdRoll/baker"
+	"github.com/jmespath/go-jmespath"
 )
 
 var ExpandJSONDesc = baker.FilterDesc{
@@ -17,7 +18,7 @@ var ExpandJSONDesc = baker.FilterDesc{
 }
 
 type ExpandJSONConfig struct {
-	Fields          map[string]string `help:"<json field -> record field> map, the rest will be ignored"`
+	Fields          map[string]string `help:"<JMESPath -> record field> map, the rest will be ignored"`
 	Source          string            `help:"record field that contains the json"`
 	TrueFalseValues []string          `help:"bind the json boolean values to correstponding strings" default:"[\"true\", \"false\"]"`
 }
@@ -34,7 +35,7 @@ type ExpandJSON struct {
 	cfg *ExpandJSONConfig
 
 	fields          []baker.FieldIndex
-	jsonKey         []string
+	jexp            []*jmespath.JMESPath
 	source          baker.FieldIndex
 	trueFalseValues [2][]byte
 
@@ -55,13 +56,17 @@ func NewExpandJSON(cfg baker.FilterParams) (baker.Filter, error) {
 	}
 	ut.source = val
 	// Fields
-	for k, v := range dcfg.Fields {
-		val, ok := cfg.FieldByName(v)
+	for j, f := range dcfg.Fields {
+		i, ok := cfg.FieldByName(f)
 		if !ok {
-			return nil, fmt.Errorf("field %s unknown, can't expand %s into it", v, k)
+			return nil, fmt.Errorf("field %q unknown, can't expand %q into it", f, j)
 		}
-		ut.fields = append(ut.fields, val)
-		ut.jsonKey = append(ut.jsonKey, k)
+		c, err := jmespath.Compile(j)
+		if err != nil {
+			return nil, fmt.Errorf("can't compile JMESPath expression %q for field %q", j, f)
+		}
+		ut.fields = append(ut.fields, i)
+		ut.jexp = append(ut.jexp, c)
 	}
 	// TrueFalseValues
 	if l := len(dcfg.TrueFalseValues); l != 2 {
@@ -82,49 +87,47 @@ func (f *ExpandJSON) Stats() baker.FilterStats {
 }
 
 func (f *ExpandJSON) Process(l baker.Record, next func(baker.Record)) {
-	// custom json decoder to get all json value types as strings
-	// really we just want string -> bytes map
-	gm := f.processJSON(l.Get(f.source))
+	data := f.processJSON(l.Get(f.source))
 
-	if gm != nil {
-		for i, k := range f.jsonKey {
-			v, ok := gm[k]
-			if !ok {
-				continue
-			}
-			l.Set(f.fields[i], v)
+	for i, c := range f.jexp {
+		r, err := c.Search(data)
+		if err != nil || r == nil {
+			continue
 		}
+		l.Set(f.fields[i], f.postProcessJSON(r))
 	}
+
 	atomic.AddInt64(&f.numProcessedLines, 1)
 	next(l)
 }
 
-func (f *ExpandJSON) processJSON(data []byte) map[string][]byte {
+func (f *ExpandJSON) processJSON(data []byte) interface{} {
 	if len(data) == 0 {
 		return nil
 	}
 	d := json.NewDecoder(bytes.NewReader(data))
-	d.UseNumber()
-	var x map[string]interface{}
+	d.UseNumber() // leave numbers as strings
+	var x interface{}
 	if err := d.Decode(&x); err != nil {
 		return nil
 	}
-	gm := make(map[string][]byte)
-	for k, v := range x {
-		switch typedValue := v.(type) {
-		case json.Number:
-			gm[k] = []byte(typedValue)
-		case string:
-			gm[k] = []byte(typedValue)
-		case bool:
-			if typedValue {
-				gm[k] = f.trueFalseValues[trueIdx]
-			} else {
-				gm[k] = f.trueFalseValues[falseIdx]
-			}
-		default:
-			// skip other values, including nested json
+	return x
+}
+
+func (f *ExpandJSON) postProcessJSON(r interface{}) (val []byte) {
+	switch typedValue := r.(type) {
+	case json.Number:
+		val = []byte(typedValue)
+	case string:
+		val = []byte(typedValue)
+	case bool:
+		if typedValue {
+			val = f.trueFalseValues[trueIdx]
+		} else {
+			val = f.trueFalseValues[falseIdx]
 		}
+	default:
+		val, _ = json.Marshal(typedValue)
 	}
-	return gm
+	return
 }