--- a/filter/expand_json_test.go
+++ b/filter/expand_json_test.go
@@ -20,43 +20,81 @@ func TestExpandJSON(t *testing.T) {
 		{
 			name: "string test1",
 			json: `{"j1": "value1", "j2": "value 2", "j3": "1,[2],3,4{"}`,
-			want: map[string]string{"f1": "value1", "f2": "value 2", "f3": "1,[2],3,4{"},
+			want: map[string]string{
+				"f1": "value1",
+				"f2": "value 2",
+				"f3": "1,[2],3,4{",
+			},
 		},
 		{
 			name: "string test2",
 			json: `{"junk1": 234, "junk2": {"a":"hello","b":[true,2]}, "j1": "text"}`,
-			want: map[string]string{"f1": "text"},
+			want: map[string]string{
+				"f1": "text",
+			},
 		},
 		{
 			name: "test numbers",
 			json: `{"j1": 12, "j2": 0.15, "j3": 15.12312312312}`,
-			want: map[string]string{"f1": "12", "f2": "0.15", "f3": "15.12312312312"},
+			want: map[string]string{
+				"f1": "12",
+				"f2": "0.15",
+				"f3": "15.12312312312",
+			},
 		},
 		{
 			name: "test boolean 1",
 			json: `{"j1": true, "j2": false}`,
-			want: map[string]string{"f1": "true", "f2": "false"},
+			want: map[string]string{
+				"f1": "true",
+				"f2": "false",
+			},
 		},
 		{
-			name:      "test boolean 2",
-			json:      `{"j1": true, "j2": false}`,
-			want:      map[string]string{"f1": "t", "f2": "f"},
+			name: "test boolean 2",
+			json: `{"j1": true, "j2": false}`,
+			want: map[string]string{
+				"f1": "t",
+				"f2": "f",
+			},
 			trueFalse: []string{"t", "f"},
 		},
 		{
 			name: "other json types",
 			json: `{"j1": null, "j2": [1,2,3,4], "j3": {"j1":"a", "j2":false}}`,
-			want: map[string]string{"f1": "", "f2": "[1,2,3,4]", "f3": `{"j1":"a","j2":false}`},
+			want: map[string]string{
+				"f1": "",
+				"f2": "[1,2,3,4]",
+				"f3": `{"j1":"a","j2":false}`,
+			},
 		},
 		{
 			name: "empty json",
 			json: ``,
-			want: map[string]string{"f1": "", "f2": ""},
+			want: map[string]string{
+				"f1": "",
+				"f2": "",
+			},
 		},
 		{
 			name: "json parse error",
 			json: `{this is not a json]`,
-			want: map[string]string{"f1": "", "f2": ""},
+			want: map[string]string{
+				"f1": "",
+				"f2": "",
+			},
+		},
+		{
+			name: "more complex JMESPath expression",
+			json: `[{"name": "name1"}, {"name": "name2"}]`,
+			field: map[string]string{
+				"[0].name": "f1",
+				"[1].name": "f2",
+			},
+			want: map[string]string{
+				"f1": "name1",
+				"f2": "name2",
+			},
 		},
 
 		// errors
@@ -72,6 +110,15 @@ func TestExpandJSON(t *testing.T) {
 			field: map[string]string{
 				"j1": "not_exist",
 				"j2": "f2",
+			},
+			wantErr: true,
+		},
+		{
+			name: "JMESPath malformed",
+			json: `{"j1": true, "j2": false}`,
+			field: map[string]string{
+				"j1":  "f1",
+				"js.": "f2",
 			},
 			wantErr: true,
 		},
@@ -106,7 +153,7 @@ func TestExpandJSON(t *testing.T) {
 					"j3": "f3",
 				}
 			}
-			if len(tt.source) == 0 {
+			if tt.source == "" {
 				tt.source = "j"
 			}
 			jsonConfig := ExpandJSONConfig{
@@ -116,6 +163,7 @@ func TestExpandJSON(t *testing.T) {
 			if len(tt.trueFalse) != 0 {
 				jsonConfig.TrueFalseValues = tt.trueFalse
 			}
+
 			cfg := baker.FilterParams{
 				ComponentParams: baker.ComponentParams{
 					FieldByName:   fieldByName,
@@ -141,11 +189,6 @@ func TestExpandJSON(t *testing.T) {
 					if !bytes.Equal(rec2.Get(i), []byte(v)) {
 						t.Errorf("got %q, want %q", rec2.Get(i), v)
 					}
-				}
-				// check that the json field is untouched
-				i, _ := fieldByName("j")
-				if !bytes.Equal(rec2.Get(i), []byte(tt.json)) {
-					t.Errorf("got %q, want %q", rec2.Get(i), tt.json)
 				}
 			})
 		})