#### :question: What

Add the `ExpandJSON` filter. The filter copy the values of a set JSON keys to corresponding record fields.
The filter support also [JMESpath](https://jmespath.org/tutorial.html) to select the values to copy inside the JSON.

A possible filter configuration is:

```
[[filter]]
name="ExpandJSON"
	[filter.config]
	Source = "json_data"

	[filter.config.Fields]
	jfield1  = "field1"
        jfield2  = "field2"
```

The filter will transform the following input in the corresponding output:

__Input:__

| field1 | field2 | json_data                                     |
|:-------:|:-------:|:-------------------------------------------:|
|          |           | {jfield1:"value1", jfield2:"value2"} |

__Output:__

| field1   | field2  | json_data                                     |
|:---------:|:--------:|:-------------------------------------------:|
| value1 | value2 | {jfield1:"value1", jfield2:"value2"} |

#### :hammer: How to test

1. Configure the filter in the TOML
2. Change the CSV delimiter to something like `;` to avoid a clash with the JSON
3. Run Baker with an appropriate input

#### :white_check_mark: Checklists

- [x] Is there unit/integration test coverage for all new and/or changed functionality added in this PR?
- [x] Have the changes in this PR been functionally tested?
- [x] Have new components been added to the related `all.go` files?
- [ ] Have new components been added to the documentation website?
- [x] Has `make gofmt-write` been run on the code?
- [x] Has `make govet` been run on the code? Has the code been fixed accordingly to the output?
- [x] Have the changes been added to the [CHANGELOG.md](./CHANGELOG.md) file?
- [x] Have the steps in [CONTRIBUTING.md](./CONTRIBUTING.md) been followed to update a Go module?