@lucarin91 2021-02-10
### UPDATE: add JMESpath support

The last version of the PR uses the library [`go-jmespath`](https://github.com/jmespath/go-jmespath) for selecting the JSON value to copy.
This solution will increase both the versatility of the filter and the performance.

Here the performance comparisons between the naive implementation with a map and the actual implementation with JMESpath:

```
name          old time/op    new time/op    delta
ExpandJSON-8    3.01µs ±10%    1.70µs ±13%  -43.52%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
ExpandJSON-8    1.97kB ± 0%    1.38kB ± 0%  -29.67%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
ExpandJSON-8      31.0 ± 0%      20.0 ± 0%  -35.48%  (p=0.000 n=10+10)
```